requirements.txt
torch
transformers
datasets
seqeval
python-dotenv
pyyaml
tqdm
Pillow
matplotlib
